+++ templates/ecommerce/src/graphql/meta.ts
@@ -1,0 +1,9 @@
+import { MEDIA_FIELDS } from './media'
+
+export const META = `meta {
+  title
+  image {
+    ${MEDIA_FIELDS}
+  }
+  description
+}`